requirements.txt
accelerate==1.2.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiosignal==1.3.1
anyio==4.7.0
attrs==24.2.0
bitsandbytes==0.41.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
Cython==0.29.37
decord==0.6.0
einops==0.6.1
einops-exts==0.0.4
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.55.2
frozenlist==1.5.0
fsspec==2024.10.0
h11==0.14.0
hjson==3.1.0
huggingface-hub==0.26.5
idna==3.10
iopath==0.1.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
linkify-it-py==2.0.3
markdown-it-py==2.2.0
markdown2==2.4.0
MarkupSafe==3.1.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
narwhals==1.16.0
networkx==3.4.2
ninja==1.11.1
numpy==1.26.4
opencv-python==4.10.0.84
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==11.0.0
portalocker==3.0.0
propcache==0.2.1
protobuf==4.23.4
psutil==6.1.0
py-cpuinfo==9.0.0
pydantic==1.10.19
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
rpds-py==0.22.3
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.14.1
semantic-version==2.10.0
sentencepiece==0.1.99
setuptools==75.1.0
shortuuid==1.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.27.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.47.0
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.2.3
uvicorn==0.22.0
websockets==14.1
wheel==0.44.0
yarl==1.18.3